requirements.txt
pypng==0.0.20
numpy==1.18.5